requirements.txt
redis>=3.0.1
toolz>=0.9.0
fire>=0.1.3
attrs>=18.2.0